site simplification
Nuked a bunch of content that was very old/not great to have. Split most of the copy into an about page with the important stuff in the home page. Removed the twitter link; the other ones might want to be removed as well. Probably should rewrite this to use grid or something.